requirements.txt
# General utilities
uv==0.5.0
fastapi==0.110.0
uvicorn==0.24.0
python-dotenv==1.0.1

# Google APIs & Auth
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-storage==2.18.0
gspread==5.10.0
oauth2client==4.1.3
google-generativeai==0.1.1

# Web scraping & automation
requests==2.32.3
pyyaml==6.0.2
selenium==4.32.0
playwright==1.49.0
beautifulsoup4==4.12.3

# Blockchain/Web3
web3==6.1.0
eth-account==0.7.0

# GitHub/API integrations
pygithub==1.58

# Testing
pytest==8.3.4

# Alpaca
alpaca-trade-api==2.1.0
alpaca-py==0.1.0

# Yahoo Finance
yfinance==0.2.2